Also moved duplicated response logic to base class
